+++ examples/modern-minimal/modern.config.ts
@@ -25,9 +25,9 @@ export default defineConfig({
       chain.plugin(pluginName).use(RsdoctorWebpackPlugin, [
         {
           disableClientServer: !process.env.ENABLE_CLIENT_SERVER,
-          features: ['bundle', 'plugins', 'loader', 'resolver']
+          features: ['bundle', 'plugins', 'loader', 'resolver'],
         },
       ]);
-    }
-  }
+    },
+  },
 });