+++ packages/types/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rsdoctor/types
 
+## 0.1.6
+
+### Patch Changes
+
+- 8527171: fix(inner-doctor): support getHash for inner rsdoctor
+- 5986abf: feat(inner-rsdoctor): add innerClientName for inner-rsdoctor
+- e322824: feat(client): add bundle diff components
+
 ## 0.1.5
 
 ## 0.1.4